+++ src/main/java/controllers/habits/TrackHabitController.java
@@ -2,7 +2,6 @@ package controllers.habits;
 
 import use_cases.habits.track_habit.TrackHabitInputBoundary;
 import use_cases.habits.track_habit.TrackHabitInputData;
-import use_cases.habits.track_habit.TrackHabitOutputData;
 
 /**
  Controller class for editing a habit.
@@ -23,10 +22,9 @@ public class TrackHabitController {
 
     /**
      * @param name - The name of habit.
-     * @return returns the output data after editing.
      */
-    public TrackHabitOutputData track(String name){
+    public void track(String name){
         TrackHabitInputData inputData = new TrackHabitInputData(name);
-        return userInput.track(inputData);
+        userInput.track(inputData);
     }
 }
