+++ src/main/java/entities/CommonUser.java
@@ -15,7 +15,7 @@ public class CommonUser implements User, Serializable{
     private String lastName;
     private String emailAddress;
     private String password;
-    private HabitFactory habitCollection;
+    private HabitCollection habitCollection;
     private CategoryCollection categoryCollection;
     private TimerTomato tomato;
 
@@ -35,7 +35,7 @@ public class CommonUser implements User, Serializable{
         this.password = password;
         this.firstName = first_name;
         this.lastName = last_name;
-        this.habitCollection = new HabitFactory();
+        this.habitCollection = new HabitCollection();
         this.categoryCollection = new CategoryCollection();
         this.tomato = new TimerTomato();
     }
@@ -85,7 +85,7 @@ public class CommonUser implements User, Serializable{
      * @return the user's collection of habits
      */
     @Override
-    public HabitFactory getHabitCollection(){
+    public HabitCollection getHabitCollection(){
         // returns the user's habit collection
         return this.habitCollection;
     }
