+++ src/main/java/entities/TaskFactory.java
@@ -1,5 +1,6 @@
 package entities;
 
+import java.io.Serializable;
 import java.util.ArrayList;
 import java.util.HashMap;
 
@@ -7,7 +8,7 @@ import java.util.HashMap;
  * Represents a collection of tasks.
  */
 
-public class TaskFactory extends Factory<Task>{
+public class TaskFactory extends Factory<Task> implements Serializable{
     private HashMap<Integer, Task> tasks;
 
     public TaskFactory(){
