--- src/main/java/ui/habits/LabelTextPanel.java
@@ -1,11 +1,0 @@
-package ui.habits;
-
-import javax.swing.*;
-
-public class LabelTextPanel extends JPanel {
-
-    public LabelTextPanel(JLabel label, JTextField textField){
-        this.add(label);
-        this.add(textField);
-    }
-}
