+++ src/main/java/use_cases/tasks/create_task/CreateTaskDsRequestModel.java
@@ -31,11 +31,6 @@ public class CreateTaskDsRequestModel {
         return this.id;
         }
 
-     * @return the completion status of the created task.
-     */
-    public boolean isCompleteStatus() {
-        return completeStatus;
-    }
 
     /**
      * @param task - Task object to be created
