+++ src/main/java/use_cases/tasks/edit_task/EditTaskDsRequestModel.java
@@ -38,10 +38,4 @@ public class EditTaskDsRequestModel {
      */
     public void setId(Integer id){this.id = id;}
 
-
-     * @return the completion status of the task.
-     */
-    public boolean isCompletionStatus() {
-        return completionStatus;
-    }
 }
