+++ src/main/java/use_cases/timer_use_cases/TimerCycle.java
@@ -1,6 +1,6 @@
 package use_cases.timer_use_cases;
 
-/*
+
 import entities.TimerTomato;
 
 import javax.swing.*;
@@ -77,4 +77,4 @@ public class TimerCycle {
 
         countDown.start();
     }
-}*/
+}
