+++ src/test/java/test_use_cases/TestCategoryUseCases.java
@@ -3,13 +3,13 @@
 package test_use_cases;
 
 import entities.Category;
-import entities.CategoryFactory;
+import entities.CategoryCollection;
 import use_cases.categories.create_category.*;
 
 import static org.junit.Assert.*;
 
 public class TestCategory {
-     CategoryFactory factory = new CategoryFactory();
+     CategoryCollection factory = new CategoryCollection();
      String name = "Chores";
      String colour = "RED";
      CreateCategoryOutputBound outputBound = new CreateCategoryOutputBound() {
@@ -26,17 +26,17 @@ public class TestCategory {
     } ;
 
     CreateCategoryInputBound inputBound = new CreateCategoryInputBound() {
-        @Override
+        //@Override
         public CreateCategoryOutputData edit(CreateCategoryInputData inputData) {
             if (inputData.getName().isBlank()) {
                 CreateCategoryOutputData outputData = new CreateCategoryOutputData("Error: Please enter the name of " +
                         "the category.");
-                return outputBound.prepareFailView(outputData);
-            } else if(factory.contains(inputData.getName(), false)){
+                //return outputBound.prepareFailView(outputData);
+            } //else if(factory.contains(inputData.getName(), false)){
                 CreateCategoryOutputData outputData = new CreateCategoryOutputData("Error: This category name " +
                         "already exists. Please enter a new category name.");
-                return outputBound.prepareFailView(outputData);
-            }
+                //return outputBound.prepareFailView(outputData);
+            //}
             Category category = new Category(inputData.getName(), inputData.getColour());
             factory.addItem(category);
             CreateCategoryOutputData outputData = new CreateCategoryOutputData(category);
